Next phase - app launches and main page displays.
