+++ src/android/stripe.ts
@@ -1,12 +1,10 @@
 import * as utils from "tns-core-modules/utils/utils";
-import * as common from "../stripe.common";
 
 declare const com;
-export class Stripe extends common.Stripe {
+export class Stripe {
     private _stripe: any /* com.stripe.android.Stripe */;
     private _ctx;
     constructor(apiKey: string) {
-        super();
         this._ctx = utils.ad.getApplicationContext();
         this._stripe = new com.stripe.android.Stripe(this._ctx, apiKey);
     }
@@ -33,5 +31,8 @@ export class Stripe extends common.Stripe {
     }
 }
 
-export class StripeCustomerContext extends common.StripeCustomerContext {
+export class StripeCustomerContext {
+}
+
+export class StripePaymentContext {
 }
