--- src/card-view.common.ts
@@ -1,7 +1,0 @@
-import { View } from "tns-core-modules/ui/core/view";
-
-export class CreditCardView extends View {
-
-}
-
-export type CardBrand = "Visa" | "Amex" | "MasterCard" | "Discover" | "JCB" | "DinersClub" | "Unknown";
