--- src/card-view.d.ts
@@ -1,11 +1,0 @@
-import { Card } from "./card";
-import * as common from "./card-view.common";
-export declare class CreditCardView extends common.CreditCardView {
-    nativeView: any;
-    createNativeView(): Object;
-    initNativeView(): void;
-    disposeNativeView(): void;
-    readonly card: Card;
-}
-
-export type CardBrand = "Visa" | "Amex" | "MasterCard" | "Discover" | "JCB" | "DinersClub" | "Unknown";
