--- src/card.common.ts
@@ -1,2 +1,0 @@
-export class Card {
-}
