--- src/card.d.ts
@@ -1,29 +1,0 @@
-import { CardBrand } from "./card-view.common";
-import * as common from "./card.common";
-export declare class Card extends common.Card {
-    private _card;
-    constructor(cardNumber: string, cardExpMonth: any, cardExpYear: any, cardCVC: string);
-    readonly card: any;
-    validateNumber(): boolean;
-    validateCVC(): boolean;
-    validateCard(): boolean;
-    validateExpMonth(): boolean;
-    validateExpYear(): boolean;
-    readonly number: string;
-    readonly cvc: string;
-    readonly expMonth: any;
-    readonly expYear: any;
-    name: string;
-    addressLine1: string;
-    addressLine2: string;
-    addressCity: string;
-    addressZip: string;
-    addressState: string;
-    addressCountry: string;
-    currency: string;
-    readonly last4: string;
-    readonly brand: CardBrand;
-    readonly fingerprint: string;
-    readonly funding: string;
-    readonly country: string;
-}
